:package: :: move package entity to mapper
